marged, default fd error remain
