--- wd_lounch/execvp_test.c
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   execvp_test.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ewatanab <ewatanab@student.42tokyo.jp>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/02/06 13:34:58 by ewatanab          #+#    #+#             */
-/*   Updated: 2021/02/06 14:19:10 by ewatanab         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include <unistd.h>
-
-int		main()
-{
-	//char *ls[] = {"ls", NULL};
-	char *grep[] = {"grep", "pipe", NULL};
-
-	execvp(grep[0], grep);
-}
